backend/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
confusable-homoglyphs==3.3.1
filelock==3.17.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Levenshtein==0.26.1
MarkupSafe==3.0.2
RapidFuzz==3.12.1
requests==2.32.3
requests-file==2.1.0
tldextract==5.1.3
urllib3==2.3.0
Werkzeug==3.1.3
python-whois==0.9.5

